graphql: schema / query generation
